+++ modules/canvaskit/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
  - Extern bug in `CanvasKit.SkMatrix.invert`
+ - Fallback to CPU now properly refreshes the canvas to get access to the
+   CanvasRenderingContext2D.
 
 ## [0.5.0] - 2019-03-08